refactor jna
