+++ interactive_engine/src/executor/Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "store",
     "server",
     "Pegasus",
+    "jna",
 ]
 
 [profile.release]
