+++ interactive_engine/src/executor/jna/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "maxgraph-jna"
+version = "0.1.0"
+authors = ["maxgraph@alibaba-inc.com"]
+
+[lib]
+crate-type = ["rlib", "cdylib"]
+
+[dependencies]
+maxgraph-runtime = { path = "../runtime" }
+maxgraph-store = { path = "../store" }
+maxgraph-common = { path = "../../common/rust/common" }
+maxgraph-server = { path = "../server" }
+pegasus = { path = "../Pegasus" }
+itertools = "0.7.8"
+log = "0.3"
+log4rs = "0.8.0"
+lazy_static = "1.2.0"
+grpcio = "=0.4.1"
