+++ interactive_engine/src/executor/jna/src/executor/pegasus/mod.rs
@@ -1,4 +1,4 @@
 mod executor;
 mod executor_server;
-mod pegasus_server_manager;
 mod jna_server_response;
+mod pegasus_server_manager;
