+++ interactive_engine/src/executor/jna/src/executor/pegasus/pegasus_server_manager.rs
@@ -1,17 +1,16 @@
-use server::manager::{ServerManager, ManagerGuards};
 use std::sync::{Arc, RwLock};
 use maxgraph_store::config::StoreConfig;
 use pegasus::{Pegasus, ConfigArgs};
-use store::task_partition_manager::TaskPartitionManager;
 use std::net::{TcpListener, TcpStream};
-use server::allocate::register_tcp_listener;
 use std::{thread, time};
 use pegasus::network::reconnect;
-use tokio::io::Error;
 use pegasus::network_connection::Connection;
 use std::sync::atomic::AtomicBool;
 use maxgraph_store::db::api::{GraphResult, GraphError};
 use maxgraph_store::db::api::GraphErrorCode::InvalidOperation;
+use maxgraph_runtime::store::task_partition_manager::TaskPartitionManager;
+use maxgraph_runtime::server::allocate::register_tcp_listener;
+use maxgraph_runtime::server::manager::ManagerGuards;
 
 pub struct PegasusServerManager {
     pegasus_runtime: Arc<Option<Pegasus>>,
