+++ interactive_engine/src/executor/jna/src/store/graph.rs
@@ -2,13 +2,13 @@
 use std::os::raw::{c_char, c_void};
 use std::ffi::CStr;
 use std::str;
-use crate::db::api::{GraphConfigBuilder, SnapshotId, GraphResult, GraphStorage, Vertex, TypeDef, EdgeId, EdgeKind, GraphError, DataLoadTarget};
-use crate::db::jna::jna_response::JnaResponse;
-use crate::db::api::PropertyMap;
-use crate::db::proto::common::{OpTypePb, OperationBatchPb, OperationPb, DataOperationPb, VertexIdPb, LabelIdPb, EdgeIdPb, EdgeKindPb, TypeDefPb, DdlOperationPb, ConfigPb, CreateVertexTypePb, AddEdgeKindPb, PrepareDataLoadPb, CommitDataLoadPb};
-use crate::db::graph::store::GraphStore;
-use crate::db::common::bytes::util::parse_pb;
+use maxgraph_store::db::api::{GraphConfigBuilder, SnapshotId, GraphResult, GraphStorage, Vertex, TypeDef, EdgeId, EdgeKind, GraphError, DataLoadTarget};
+use maxgraph_store::db::api::PropertyMap;
+use maxgraph_store::db::proto::common::{OpTypePb, OperationBatchPb, OperationPb, DataOperationPb, VertexIdPb, LabelIdPb, EdgeIdPb, EdgeKindPb, TypeDefPb, DdlOperationPb, ConfigPb, CreateVertexTypePb, AddEdgeKindPb, PrepareDataLoadPb, CommitDataLoadPb};
+use maxgraph_store::db::graph::store::GraphStore;
+use maxgraph_store::db::common::bytes::util::parse_pb;
 use std::sync::Once;
+use store::jna_response::JnaResponse;
 
 pub type GraphHandle = *const c_void;
