--- interactive_engine/src/executor/store/src/db/jna/mod.rs
@@ -1,2 +1,0 @@
-mod graph;
-mod jna_response;
