+++ interactive_engine/src/executor/store/src/db/mod.rs
@@ -6,4 +6,3 @@ pub mod graph;
 mod util;
 #[allow(bare_trait_objects)]
 pub mod proto;
-mod jna;
